update INT_MIN and INT_MAX

+++ main.h
@@ -1,5 +1,5 @@
 #ifndef MAIN_H
 #define MAIN_H
 int _printf(const char *format, ...);
-int _printint(int n);
+int _printint(long n);
 #endif /* MAIN_H */
